feat(HIGH-27): shorts hook 분리
